Fix all cargo clippy warnings in arm7tdmi and core crates
Agent-Logs-Url: https://github.com/michelhe/rustboyadvance-ng/sessions/4c1738e3-5bd2-433f-97cb-5fecbc4e377f

+++ core/src/gpu/render/text.rs
@@ -37,7 +37,7 @@ impl Gpu {
             (256, 512) => bg_y / 256,
             (512, 512) => index2d!(u32, bg_x / 256, bg_y / 256, 2),
             _ => unreachable!(),
-        } as u32;
+        };
 
         let mut se_row = (bg_x / 8) % 32;
         let se_column = (bg_y / 8) % 32;
